ARCPOC-572 Refactor app code functionality (#109)
* Convereted app code to open api paging mechanism, as well as use MapStruct
* Add some logging
* Finish off app code conversion
* Add some java doc
* Force build
* ARCPOC-647
* Adjusted entities to align with flyway scripts

* Convert missed entities

* Fix test

* Correct spotless problem

* - Respond to the code review comments and add more testing.

* - Cast code and title to String in the JPQL to avoid exceptions being thrown from the data layer.
- Revert the Duplicate result exception change.
- Refactor unit tests.

---------

+++ src/testCommon/java/uk.gov.hmcts.appregister/data/FeeTestData.java
@@ -2,6 +2,7 @@ package uk.gov.hmcts.appregister.data;
 
 import static org.instancio.Select.field;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 import org.instancio.Instancio;
@@ -19,7 +20,7 @@ public class FeeTestData
         data.reference(StringUtil.stripToMax(uniqueId.toString(), 12))
                 .description("description" + uniqueId)
                 .startDate(LocalDate.now())
-                .amount(20D)
+                .amount(BigDecimal.valueOf(20))
                 .build();
 
         return data;
